Refactor experience systems and storage

+++ src/main/kotlin/db/repository/UserStatsTable.kt
@@ -4,5 +4,5 @@ import org.jetbrains.exposed.sql.Table
 
 object UserStatsTable : Table() {
     val userId = reference("user_id", UserTable)
-    val miningExp = integer("mining_exp").default(0)
+    val miningExp = long("mining_exp").default(0)
 }
