+++ src/main/kotlin/domain/model/LevelProgress.kt
@@ -10,19 +10,19 @@ package com.bitwiserain.pbbg.domain.model
  */
 data class LevelProgress(
     val level: Int,
-    val absoluteExp: Int,
-    private val absoluteExpCurrentLevel: Int,
-    private val absoluteExpNextLevel: Int
+    val absoluteExp: Long,
+    private val absoluteExpCurrentLevel: Long,
+    private val absoluteExpNextLevel: Long
 ) {
     /**
      * The current experience amount relative the previous level up threshold.
      */
-    val relativeExp: Int
+    val relativeExp: Long
         get() = absoluteExp - absoluteExpCurrentLevel
 
     /**
      * The experience to the next level up relative to the previous level up threshold.
      */
-    val relativeExpNextLevel: Int
+    val relativeExpNextLevel: Long
         get() = absoluteExpNextLevel - absoluteExpCurrentLevel
 }
