+++ src/main/kotlin/domain/model/UserStats.kt
@@ -1,5 +1,5 @@
 package com.bitwiserain.pbbg.domain.model
 
 class UserStats(
-    val miningExp: Int
+    val miningExp: Long
 )
